requirements.txt
asgiref==3.7.2
astroid==3.0.1
black==23.11.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
Deprecated==1.2.14
dill==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
isort==5.12.0
jwcrypto==1.5.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomlkit==0.12.3
urllib3==2.1.0
wrapt==1.16.0
